Note dependency on matplotlib, Python2 requirement
I briefly described the incompatibility with Python3
in GitHub issue #1

I see from the code
that the dependency on matplotlib can be avoided
with the --no_plot option
but I did not actually notice that
until a run that I did failed.
Therefore it seems better
to simply list matplotlib as a hard dependency.